python/mcpgo/requirements.txt
Django==3.1.1
djangorestframework
Pillow
django-truncate
openpyxl
mysql
